add uuid

+++ internal/service/transport/gRPC/items.go
@@ -21,7 +21,7 @@ func (s *Server) AddNewItem(ctx context.Context, in *pb.Item) (*pb.ItemID, error
 	}
 
 	return &pb.ItemID{
-		ID: int64(id),
+		ID: id,
 	}, nil
 }
 
@@ -39,7 +39,7 @@ func (s *Server) GetAllItems(ctx context.Context, in *pb.NoneObjectRequest) (*pb
 
 func (s *Server) GetItem(ctx context.Context, in *pb.ItemID) (*pb.EncodeItemResponse, error) {
 
-	res, err := s.storage.GetItem(ctx, int(in.ID))
+	res, err := s.storage.GetItem(ctx, in.ID)
 	if err != nil {
 		return nil, err
 	}
@@ -51,7 +51,7 @@ func (s *Server) GetItem(ctx context.Context, in *pb.ItemID) (*pb.EncodeItemResp
 
 func (s *Server) DeleteItem(ctx context.Context, in *pb.ItemID) (*pb.NoneObjectResp, error) {
 
-	done, err := s.storage.DeleteItem(ctx, int(in.ID))
+	done, err := s.storage.DeleteItem(ctx, in.ID)
 	if err != nil {
 		return nil, err
 	}
@@ -62,8 +62,7 @@ func (s *Server) DeleteItem(ctx context.Context, in *pb.ItemID) (*pb.NoneObjectR
 }
 
 func (s *Server) UpdateItem(ctx context.Context, in *pb.ItemID) (*pb.EncodeItemResponse, error) {
-
-	res, err := s.storage.UpdateItem(ctx, int(in.ID))
+	res, err := s.storage.UpdateItem(ctx, in.ID)
 	if err != nil {
 		return nil, err
 	}
